+++ MySession/src/test/java/jarvis/pages/FileObj.java
@@ -1,9 +1,9 @@
-package jarvis.objects;
+package jarvis.pages;
 
 
 import org.openqa.selenium.support.PageFactory;
 
-import commands.FilesOperation;
+import commands.FilesActions;
 import configs.Base;
 
 public class FileObj extends Base{
@@ -15,10 +15,10 @@ public class FileObj extends Base{
 	
 	
 	public void fileCreate(String path) {
-		FilesOperation.Create_File( getDriver(), path);
+		FilesActions.Create_File( getDriver(), path);
 	}
 	public void fileDelete(String path) {
-		FilesOperation.Delete_File( getDriver(), path);
+		FilesActions.Delete_File( getDriver(), path);
 	}
